+++ builder_pkgs/mockito/CHANGELOG.md
@@ -3,6 +3,10 @@
 * Use an empty list with a correct type argument for a fallback value for a
   method which returns Iterable.
   [#445](https://github.com/dart-lang/mockito/issues/445)
+* When selecting the library that should be imported in order to reference a
+  type, prefer a library which exports the library in which the type is
+  declared. This avoids some confusion with conditional exports.
+  [#443](https://github.com/dart-lang/mockito/issues/443)
 * Support analyzer 2.0.0
 
 ## 5.0.11